refactor connections provider detail pane
